% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who-graphs.R
\name{degrees_below_18_score}
\alias{degrees_below_18_score}
\title{degrees_below_18_score}
\usage{
degrees_below_18_score(observations)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle.observations} and further filtered by \code{remove_excluded_devices}}
}
\description{
Calculate the degrees below 18C score
}
