% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamo.R
\name{dynamo_bulk_put}
\alias{dynamo_bulk_put}
\title{dynamo_bulk_put}
\usage{
dynamo_bulk_put(.con, .table, .df, .list = NULL)
}
\arguments{
\item{.con}{a connection to a DynamoDB database (see ?paws::dynamodb)}

\item{.table}{the name of the table to append or replace rows}

\item{.df}{a data frame we wish to load}
}
\description{
Load a data frame to AWS DynamoDB. Items with matching partition key
will be overwritten
}
\examples{
# prepare a version of iris without factors and an ID column
# to match the partition key in the remote table
iris_to_put <- iris
iris_to_put$Species <- as.character(iris_to_put$Species)
iris_to_put$ID <- 300:151
\dontrun{
con <- dynamo_connect()
dynamo_bulk_put(.con = con, .table = "Iris", .df = iris_to_put)
}
}
