% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamo.R
\name{dynamo_connect}
\alias{dynamo_connect}
\title{dynamo_connect}
\usage{
dynamo_connect()
}
\description{
connect to an AWS DynamoDB given credentials stored in \code{.Renviron}.
Credentials are fetched with \code{Sys.getenv}. Required entries in the \code{.Renviron} are:
\code{ACCESS_KEY_ID}, \code{SECRET_ACCESS_KEY}, \code{PROFILE}, and \code{REGION}
}
\examples{
con <- dynamo_connect()
}
