% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamo-query.R
\name{dynamo_query}
\alias{dynamo_query}
\title{dynamo_query}
\usage{
dynamo_query(.con, .table, .partition_key, .partition_value)
}
\arguments{
\item{.con}{a connection to a DynamoDB database (see ?paws::dynamodb)}

\item{.table}{the name of the table to append or replace rows}

\item{.partition_key}{the name of the \verb{partition key} field}

\item{.partition_value}{the value of the \verb{partition key}}
}
\description{
Get data from table with a partition key. Not used.
}
\examples{
dynamo_query(
.con = dynamo_connect(),
.table = "sn-v1-devices",
.partition_key = "device_owner",
.partition_value = "al@monkeytronics.co.nz"
)
dynamo_query(
.con = dynamo_connect(),
.table = "sn-v1-interventions",
.partition_key = "device_id",
.partition_value = "W000011"
)
dynamo_query(
.con = dynamo_connect(),
.table = "sn-v1-weather-db",
.partition_key = "city",
.partition_value = "Wellington"
)
dynamo_query(
.con = dynamo_connect(),
.table = "SensorNodeData",
.partition_key = "device_id",
.partition_value = "W000011"
)
}
