% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_data_volume}
\alias{get_data_volume}
\title{get_data_volume}
\usage{
get_data_volume(observations, from_timestamp, to_timestamp)
}
\arguments{
\item{observations}{data.frame, the output of \code{wrange_observations}}

\item{from_timetamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
Check data volume and ignore devices with < 10\% data
}
\examples{
wrangled_devices <- wrangle_devices(test_params()$filtered_devices)
wrangled_weather <- wrangle_weather(
  filtered_weather = test_params()$filtered_weather,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp
)
wrangled_obs <- wrangle_observations(
  filtered_obs = test_params()$filtered_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp,
  devices = wrangled_devices,
  weather = wrangled_weather
)
data_volume <- get_data_volume(
  observations = wrangled_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp
  )
}
