% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{get_devices_for_map}
\alias{get_devices_for_map}
\title{get_devices_for_map}
\usage{
get_devices_for_map(devices, data_volume)
}
\arguments{
\item{devices}{a wrangled devices data frame as output by \code{wrangle_devices}}

\item{data_volume}{the data volume data frame as output by \code{get_data_volume}}
}
\description{
Filter the wrangled devices dataset to keep only devices with sufficient
volume for the map. This step has to go before exclusions so disconnected devices show.
}
