% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-h-graphs.R
\name{get_values_for_boxes}
\alias{get_values_for_boxes}
\title{get_values_for_boxes}
\usage{
get_values_for_boxes(observations, target_variable)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle_observations} and further filtered by \code{remove_excluded_devices}}

\item{target_variable}{character, the name of the column to use for summary}
}
\description{
get a list of summary stats for value boxes
}
