% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who-graphs.R
\name{get_who_totals}
\alias{get_who_totals}
\title{get_who_totals}
\usage{
get_who_totals(who_threshold_data)
}
\arguments{
\item{who_threshold_data}{data.frame, the output from \code{?who_threshold_data()}}
}
\description{
Totals for Value Boxes
}
