% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-home-rmd.R
\name{knit_home_report}
\alias{knit_home_report}
\title{Compile the home report}
\usage{
knit_home_report(
  param_list,
  output_file = paste0("home-", as.numeric(Sys.time()), ".html"),
  output_dir = getwd(),
  ...
)
}
\arguments{
\item{param_list}{named list of parameters, see \code{?make_param_list}}

\item{output_file}{name of the output html (default: home-{timestamp}.html)}

\item{output_dir}{path to folder to save the html (default: current work dir)}

\item{...}{additional named arguments passed onto \code{rmarkdown::render}}
}
\description{
Create a the home report
}
\examples{
obs  <- system.file("data/filtered_obs.csv", package="monkeytronics.reports")
dev  <- system.file("data/filtered_devices.csv", package="monkeytronics.reports")
wea  <- system.file("data/filtered_weather.csv", package="monkeytronics.reports")
int  <- system.file("data/filtered_interventions.csv", package="monkeytronics.reports")
from <- 1627639151
to   <- 1630317551

par_list <- make_params_list(
  filtered_obs = obs,
  filtered_devices = dev,
  filtered_weather = wea,
  filtered_interventions = int,
  fromTimeStamp = from,
  toTimeStamp = to
)
knit_home_report(par_list)

\dontrun{
# using the database
# assumes DB credentials are set as environmental variables
# see `?dynamo_connect`
dev <- dynamo_query(
.con = dynamo_connect(),
.table = "sn-v1-devices",
.partition_key = "device_owner",
.partition_value = "al@monkeytronics.co.nz"
)

int <- dynamo_query(
  .con = dynamo_connect(),
  .table = "sn-v1-interventions",
  .partition_key = "device_id",
  .partition_value = "W000011"
)

wea <- dynamo_query(
  .con = dynamo_connect(),
  .table = "sn-v1-weather-db",
  .partition_key = "city",
  .partition_value = "Wellington"
)

obs <- dynamo_query(
  .con = dynamo_connect(),
  .table = "SensorNodeData",
  .partition_key = "device_id",
  .partition_value = "W000011"
)
from <- 1627639151
to   <- 1630317551

par_list <- make_params_list(
  filtered_obs = obs,
  filtered_devices = dev,
  filtered_weather = wea,
  filtered_interventions = int,
  fromTimeStamp = from,
  toTimeStamp = to
)
knit_home_report(par_list)

}

}
