% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{make_green_area}
\alias{make_green_area}
\title{make_green_area}
\usage{
make_green_area(observations)
}
\arguments{
\item{observations}{data.frame, the output of \code{wrange_observations}}
}
\description{
Make little Green area for plot
}
