% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{make_nz_datetime}
\alias{make_nz_datetime}
\title{Unix timestamp to NZ timezone datetime}
\usage{
make_nz_datetime(timestamp)
}
\arguments{
\item{timestamp}{vector of Unix time stamps}
}
\value{
Returns "POSIXct", "POSIXt" vector in NZ time zone
}
\description{
Convert a time stamp into a NZ timezone datetime
}
\examples{
make_nz_datetime(1555555555)
}
