% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit-home-rmd.R
\name{make_params_list}
\alias{make_params_list}
\title{Param list}
\usage{
make_params_list(.list = NULL, ...)
}
\arguments{
\item{.list}{A named list of key-value pairs}

\item{...}{Unquoted key-value pairs}
}
\description{
Create a list of parameters for knitting rmd
}
\details{
Field options are:
filtered_obs,
filtered_weather,
filtered_devices,
filtered_interventions,
report_params,
fromTimeStamp,
toTimeStamp
}
\examples{
# convenient for interactive typing
create_param_list(filtered_obs = "inst/data/filtered_obs.csv")
# convenient for programming
create_param_list(.list = list(filtered_obs = "inst/data/filtered_obs.csv"))
}
