% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{nrow_values}
\alias{nrow_values}
\title{misc_values}
\usage{
nrow_values(devices, observations)
}
\arguments{
\item{devices}{a wrangled devices data frame as output by \code{wrangle_devices}}

\item{observations}{data.frame, the output of \code{wrange_observations}}
}
\description{
calculate number of rows of some datasets
}
