% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{remove_excluded_devices}
\alias{remove_excluded_devices}
\title{remove_excluded_devices}
\usage{
remove_excluded_devices(target_data, data_volume)
}
\arguments{
\item{target_data}{data.frame, either the devices or observations tables}

\item{data_volume}{the data volume data frame as output by \code{get_data_volume}}
}
\description{
Remove rows from \code{target_data} if they are flagged to \code{exclude}
in the data_volume calculation
}
