% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{report_period}
\alias{report_period}
\title{Print a formatted from-to sentence}
\usage{
report_period(from_timestamp, to_timestamp)
}
\arguments{
\item{from_timetamp, to_timestamp}{timestamps for the begin and end of the report}
}
\value{
Character
}
\description{
Takes 2 time stapms and prints a formatted from-to sentence. Doesn't check if the 2nd timestamp is greater than the 1st one.
}
\examples{
report_period(1555555555, 1666666666)
}
