% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-h-graphs.R
\name{set_options}
\alias{set_options}
\title{set_options}
\usage{
set_options(observations)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle_observations} and further filtered by \code{remove_excluded_devices}}
}
\description{
Set options for downstream charts based on available data
}
