% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-params.R
\name{test_params}
\alias{test_params}
\title{Monkeytronics default parameters}
\usage{
test_params()
}
\value{
list with default settings
}
\description{
Returns a list with links to the stored dataets
}
\examples{
test_params()
}
