% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-h-graphs.R
\name{ts_chart}
\alias{ts_chart}
\title{ts_chart}
\usage{
ts_chart(observations, from_timestamp, to_timestamp, target_variable)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle.observations} and further filtered by \code{remove_excluded_devices}}

\item{target_variable}{character, the name of the column to use for plotting
Options are (for now): \code{temp}, \code{hum}}

\item{from_timetamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
make a ggplotly time series chart for temperature, CO2, ...
}
\examples{
wrangled_devices <- wrangle_devices(test_params()$filtered_devices)
wrangled_weather <- wrangle_weather(
  filtered_weather = test_params()$filtered_weather,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp
)
wrangled_obs <- wrangle_observations(
  filtered_obs = test_params()$filtered_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp,
  devices = wrangled_devices,
  weather = wrangled_weather
)
data_volume <- get_data_volume(
  observations = wrangled_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp
  )
wrangled_devices <- remove_excluded_devices(
  target_data = wrangled_devices,
  data_volume = data_volume
)
wrangled_obs <- remove_excluded_devices(
  target_data = wrangled_obs,
  data_volume = data_volume
)
ts_chart(
  observations = wrangled_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp,
  target_variable = "temp"
)
ts_chart(
  observations = wrangled_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp,
  target_variable = "hum"
)
}
