% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who-graphs.R
\name{who_threshold_chart}
\alias{who_threshold_chart}
\title{who_threshold_chart}
\usage{
who_threshold_chart(who_threshold_data, from_timestamp, to_timestamp)
}
\arguments{
\item{who_threshold_data}{data.frame, the output from \code{?who_threshold_data()}}

\item{from_timestamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
WHO threshold chart
}
