% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who-graphs.R
\name{who_threshold_data}
\alias{who_threshold_data}
\title{who_threshold_data}
\usage{
who_threshold_data(observations)
}
\arguments{
\item{observations}{data.frame, the wrangled observations data frame as output by
\code{wrangle.observations} and further filtered by \code{remove_excluded_devices}}
}
\description{
Arrange the measurement according to WHO guidalines
}
