% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_devices}
\alias{wrangle_devices}
\title{wrangle_devices}
\usage{
wrangle_devices(
  filtered_devices,
  data_type_filter = c("CD", "TH"),
  na.strings = c("", NA),
  integer_columns = c("floor"),
  double_columns = c("long", "lat"),
  character_columns = c("device_owner", "device_id", "data_type", "comms_type",
    "locationCode", "addr", "suburb", "city", "country", "room", "hhi", "cluster",
    "privacy", "risk", "buildingType", "tenure", "height", "direction", "school",
    "building", "classRoom", "deployment_id")
)
}
\arguments{
\item{filtered_devices}{a data.frame or path to file}
}
\description{
wrangle the devices data
}
\examples{
wrangle_devices(test_params()$filtered_devices)
}
