% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_observations}
\alias{wrangle_observations}
\title{wrangle_observations}
\usage{
wrangle_observations(
  filtered_obs,
  from_timestamp,
  to_timestamp,
  devices,
  weather
)
}
\arguments{
\item{filtered_obs}{a data.frame or path to file}

\item{devices}{a wrangled devices data frame as output by \code{wrangle_devices}}

\item{weather}{a wrangled outdoor weather data frame as output by \code{wrangle_weather}}

\item{from_timetamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
wrangle the measurements
}
\examples{
wrangled_devices <- wrangle_devices(test_params()$filtered_devices)
wrangled_weather <- wrangle_weather(
  filtered_weather = test_params()$filtered_weather,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp
)
wrangle_observations(
  filtered_obs = test_params()$filtered_obs,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp,
  devices = wrangled_devices,
  weather = wrangled_weather
)
}
