% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_weather}
\alias{wrangle_weather}
\title{wrangle_weather}
\usage{
wrangle_weather(filtered_weather, from_timestamp, to_timestamp)
}
\arguments{
\item{filtered_devices}{a data.frame or path to file}

\item{from_timetamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
wrangle the weather data
}
\examples{
wrangle_weather(
  filtered_weather = test_params()$filtered_weather,
  from_timestamp = test_params()$fromTimeStamp,
  to_timestamp = test_params()$toTimeStamp
)
}
